Move authenticator from ui to core
