+++ core/auth/auth_test.go
